Fritzing files for circuits
the basic circuit, controlling two relays, and a DS18B20 temperature
sensor, is made of an ATMega328p without external crystal (8Mhz).
The mains lighting is represented by two leds.
The serial communication with the Raspberry Pi is made with two wires.
The Arduino Tx is connected to a voltage divider, and goes to the
Raspberry Pi in order to lower the 5V signal.