full feature prototype files
This commit includes the fritzing files and pictures (png, jpg, svg and
pdf) for the shematics of the prototype with all the base features. This
includes cooling (via fans and/or TEC), lights control (via 2 mains
relays), water temperature monitoring for 2 tanks and air temperature.
More features may be added later on, but this is the set of features
that will be version 1.0 of the system once fully tested.