new files (v2) and php files
A new version of the files, using the usb serial. Old files are in v1,
new in v2. There are scripts to turn the lights on and off, the fan on
and off, read the 2 water temperature sensors, the air temperature
sensor,
and also the script that prints the temperatures on the LCD, aswell as
the script that auto turns the lights on and off.
It is far from perfect, as it is in early stage. It however works
reliably.
the www contains php files used to control rlieh from a web interface.
it works, but sometime it's required to refresh the page to get it to
work. It is more a proof of concept than a well polished interface.