rlieh3.ino : arduino file for the USB controller
This file contains the full code for the Rlieh USB board. Upload this to
whatever arduino that has enough GPIO and memory (uno, leonardo, mini,
etc. Basically a atmega328p, a 32u4 or more powerfull), and the heart of
rlieh is good to go. Now it can listen to serial commands through the
USB port and return xml-like results.