improvement on the second layer+printable buttons
this version improves the second level of the box, cleans the code in some places, and adds printable button that can contain an RGB LED for backlit buttons near the LCD. The PCB support has been improved with larger screws support, meant for 3mm screws (it works even with the smaller holes of the PCB). There are 2 versions of the pi zero support, for if it must be on the top or the bottom of the box.